+++ examples/simple.py
@@ -1,10 +1,10 @@
 import tensorflow as tf
 import tf_big
 
-x = tf_big.Tensor([[1,2,3,4]])
-y = tf_big.Tensor([[1,2,3,4]])
+x = tf_big.constant([[1,2,3,4]])
+y = tf_big.constant([[1,2,3,4]])
 z = x + y
 
 with tf.Session() as sess:
-  res = z.eval(session=sess)
+  res = sess.run(z)
   print(res)
