+++ setup.py
@@ -19,10 +19,11 @@ setuptools.setup(
     package_data={'tf_big': []},
     python_requires=">=3.5",
     install_requires=[
+        "numpy >=1.14",
         "tensorflow >=1.13.1, <2",
     ],
     extras_require={
-        "tf": ["tensorflow>=1.31.1,<2"],
+        "tf": ["tensorflow >=1.13.1, <2"],
     },
     license="Apache License 2.0",
     url="https://github.com/tf-encrypted/tf-big",
