--- tf_big/python/ops/__init__.py
@@ -1,11 +1,0 @@
-# from __future__ import absolute_import
-
-# from .big_ops import big_import
-# from .big_ops import big_export
-# from .big_ops import big_add
-
-# __all__ = [
-#     "big_import",
-#     "big_export",
-#     "big_add",
-# ]
